+++ auctioning_platform/webapp/src/main.rs
@@ -1,3 +1,4 @@
-fn main() {
+#[tokio::main]
+async fn main() {
     println!("Hello, world!");
 }
