+++ auctioning_platform/webapp/tests/api/conftest.rs
@@ -1,17 +1,15 @@
-use std::net::{SocketAddr, TcpListener};
+use tokio::net::TcpListener;
 
 pub struct Context {
     pub address: String,
 }
 
 pub async fn spawn_app() -> Context {
-    let listener = TcpListener::bind("0.0.0.0:0".parse::<SocketAddr>().unwrap()).unwrap();
+    let listener = TcpListener::bind("0.0.0.0:0").await.unwrap();
     let addr = listener.local_addr().unwrap();
 
     tokio::spawn(async move {
-        axum::Server::from_tcp(listener)
-            .unwrap()
-            .serve(webapp::startup::create_app().into_make_service())
+        axum::serve(listener,webapp::startup::create_app())
             .await
             .unwrap();
     });
